+++ alps-core/src/main/java/org/alps/core/socket/netty/AlpsMessageDecoder.java
@@ -23,7 +23,7 @@ public class AlpsMessageDecoder extends MessageToMessageDecoder<AlpsProtocol.Alp
 
     @Override
     protected void decode(ChannelHandlerContext channelHandlerContext, AlpsProtocol.AlpsPacket alpsPacket, List<Object> list) throws Exception {
-        var packet = new AlpsPacket(alpsPacket.getConnectTypeValue(), (short) alpsPacket.getModule(), dataCoderFactory,
+        var packet = new AlpsPacket(alpsPacket.getConnectTypeValue(), alpsPacket.getModule(), dataCoderFactory,
                 buildMetadata(alpsPacket.getMetadata()), buildData(alpsPacket.getData()), alpsPacket);
         list.add(packet);
     }
